ready for launch

+++ routes/api.php
@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Http\Request;
+use App\Http\Resources\UserResource;
 
 /*
 |--------------------------------------------------------------------------
@@ -17,8 +18,10 @@ Route::group(['middleware' => 'auth:api'], function () {
     Route::post('logout', 'Auth\LoginController@logout');
 
     Route::get('/user', function (Request $request) {
-        return $request->user();
+        return new UserResource($request->user());
     });
+
+    Route::get('users', 'UserController@index');
 
     Route::patch('settings/profile', 'Settings\ProfileController@update');
     Route::patch('settings/password', 'Settings\PasswordController@update');
@@ -36,17 +39,24 @@ Route::group(['middleware' => 'guest:api'], function () {
 
 Route::group(['prefix' => '/categories'], function () {
     Route::get('', 'CategoryController@index');
-    Route::post('', 'CategoryController@store');
+    Route::post('', 'CategoryController@store')->middleware(['auth', 'admin']);
     Route::get('/{category}', 'CategoryController@show');
-    Route::patch('/{category}', 'CategoryController@update');
-    Route::delete('/{category}', 'CategoryController@destroy');
+    Route::patch('/{category}', 'CategoryController@update')->middleware(['auth', 'admin']);
+    Route::delete('/{category}', 'CategoryController@destroy')->middleware(['auth', 'admin']);
 });
 
 Route::group(['prefix' => '/events'], function () {
     Route::get('', 'EventController@index');
-    Route::post('', 'EventController@store');
+    Route::post('', 'EventController@store')->middleware('auth');
     Route::get('/{event}', 'EventController@show');
-    Route::post('/{event}', 'EventController@update');
-    Route::patch('/{event}/approve', 'EventController@approve');
-    Route::delete('/{event}', 'EventController@destroy');
+    Route::post('/{event}', 'EventController@update')->middleware('auth');
+    Route::patch('/{event}/approve', 'EventController@approve')->middleware(['auth', 'approver']);
+    Route::delete('/{event}', 'EventController@destroy')->middleware(['auth', 'approver']);
+});
+
+
+Route::group(['middleware' => ['auth:api', 'admin'], 'prefix' => '/privileges'], function () {
+    Route::get('', 'PrivilegeController@index');
+    Route::post('', 'PrivilegeController@store');
+    Route::delete('{privilege}', 'PrivilegeController@destroy');
 });
